update file names, .env variables, and Bearer Token logic
